gmmk-pro-v1: drop alt-like 65% macOS configuration
Number row and right-side buttons mirror the layout of my
Drop ALT 65% keyboard, bottom row as well follows the
macOS layout.  RGB heatmap is enabled as the default
with Fn-R being the toggle to switch modes and
Fn-Print Screen to toggle RGB on/off.  Also made the Fn-`
be Escape, like my 65% row.

Signed-off-by: B. Thomas Goodwin <bt.goodwin@me.com>